### 2022-02-16 #4177 @whedon
[PRE REVIEW]: ospgrillage: A bridge deck grillage analysis preprocessor for `OpenSeesPy`
**Submitting author:** @ccaprani (<a href="http://orcid.org/0000-0001-6166-0895">Colin C. Caprani</a>)
**Repository:** https://github.com/MonashSmartStructures/ospgrillage
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
