### 2022-02-17 #4177 @ccaprani
@danielskatz thank you for your time on this. I've fixed the doi problem.

Suggested reviewer github handles (without @) are:

**Existing JOSS**:

- iammix
- arunmano121
- vsangelidakis
- Jesusbill

**New to JOSS**:

- m-clare: OSS-oriented structural engineer
- zhuminjie: developer of OpenSeesPy
- mhscott: maintainer of OpenSees
- hackl: OSS-oriented structural engineer
- robbievanleeuwen: OSS-oriented structural engineer
- connorferster: OSS-oriented structural engineer
- fmckenna: original developer of OpenSees
