+++ ospgrillage/osp_grillage.py
@@ -124,7 +124,13 @@ class OspGrillage:
         * ext_to_int_dist: (Int or Float, or a List of Int or Float) distance between internal beams
          and exterior main beams. If list is provided (usually size 2), apply each distinct distance to left and right
          side respectively.
-
+        * multi_span_dist_list: (List of Int/Float) List of distance (x dir) correspond to span length of each multi span
+        * multi_span_num_points: (List of Int) Num of transverse member correspond to spans of each element in multi_span_dist_list
+         If not specified, takes int var for num_trans_beam and assigns to all spans of multi_span_dist_list
+        * continuous: (Bool) To set continuity of spans. Default True. If False, separate spans by non_cont_spacing_x
+        * stitch_slab_elements: (Bool) To set stictch elements between spans. Elements are set using `set_member()` with
+         member= "stich_elements"
+        * non_cont_spacing_x: (float) sets spacing or length of stitch elements.
 
         :raises ValueError: If skew angle is greater than 90. If number of transverse grid line is less than 2.
 
@@ -197,6 +203,7 @@ class OspGrillage:
             "exterior_main_beam_2",
             "start_edge",
             "end_edge",
+            "stitch_elements",
             "transverse_slab",
         ]
         # prefix index of members after longitudinal members
@@ -294,6 +301,13 @@ class OspGrillage:
         # create dict of standard elements from the generated Mesh obj
         self._create_standard_element_list()  # base class method, concrete classes may overwrite this
 
+        # flag for OpenSees model instance
+        self.model_instance = False  # default false before creating
+
+        # list storing all commands and str
+        self.model_command_list = []  # to be populated
+        self.analysis_command = None
+
     def _create_mesh(self, **kwargs):
         """
         Private function to create mesh. Creates the concrete Mesh class based on mesh type specified
@@ -335,7 +349,7 @@ class OspGrillage:
                 now = datetime.now()
                 dt_string = now.strftime("%d/%m/%Y %H:%M:%S")
                 file_handle.write("# Constructed on:{}\n".format(dt_string))
-                # write imports
+                # necessary imports
                 file_handle.write(
                     "import numpy as np\nimport math\nimport openseespy.opensees as ops"
                     "\nimport openseespy.postprocessing.Get_Rendering as opsplt\n"
@@ -351,9 +365,9 @@ class OspGrillage:
     # function to run mesh generation
     def _run_mesh_generation(self):
         """
-        Private function to write / execute commands. This function communicates with OpenSees framework
+        Private function to write / execute commands. This function handles relevant OpenSeesPy commands
         """
-        # 2 generate command lines in output py file
+        # write / execute model commands
         self._write_op_node(self.Mesh_obj)  # write node() commands
         self._write_op_fix(self.Mesh_obj)  # write fix() command for support nodes
         self._write_geom_transf(self.Mesh_obj)  # x dir members
@@ -365,6 +379,8 @@ class OspGrillage:
                     file_handle.write(mat_str)
             else:
                 eval(mat_str)
+                self.model_command_list.append(mat_str)
+
         for sec_str in self.section_command_list:
             if self.pyfile:
                 with open(self.filename, "a") as file_handle:
@@ -372,6 +388,7 @@ class OspGrillage:
                     file_handle.write(sec_str)
             else:
                 eval(sec_str)
+                self.model_command_list.append(sec_str)
 
         # write /execute element commands
         for ele_dict in self.element_command_list:
@@ -382,10 +399,14 @@ class OspGrillage:
                             file_handle.write(ele_str)
                     else:
                         eval(ele_str)
+                        self.model_command_list.append(ele_str)
+        # if created OpenSees instance, set instance flag
[…]
