+++ ospgrillage/mesh.py
@@ -7,6 +7,8 @@ by OspGrillage class.
 """
 import math
 
+import numpy as np
+
 from ospgrillage.static import *
 from collections import namedtuple
 
@@ -107,21 +109,22 @@ class Mesh:
         self.skew_threshold = [11, 30]
         self.curve = False
         self.search_x_inc = 0.001
+        self.y_elevation = 0  # plane position of grillage
         # initiate list for nodes and elements
         self.long_ele = []
         self.trans_ele = []
         self.edge_span_ele = []
-        self.y_elevation = 0
-        self.mesh_design_line = []
+        self.connect_ele = []
         # dict for node and ele transform
         self.transform_dict = dict()  # key: vector xz, val: transform tag
         self.node_spec = (
             dict()
         )  # key: node tag, val: dict of node details - see technical notes
         # variables for curve mesh
+        # if multiple curve centers are presented, assign each curve to respective spans of multi_span_dist_list
         self.curve_center = []
         self.curve_radius = []
-        # line / circle equation variables - instantiate
+        # init line / circle equation variables
         self.m = 0
         self.c = 0
         self.r = 0
@@ -132,21 +135,63 @@ class Mesh:
             self.mesh_origin = [0, 0, 0]
         else:
             self.mesh_origin = mesh_origin  # default origin
-        # meshing variables
+        # init meshing variables
         self.first_connecting_region_nodes = []
         self.end_connecting_region_nodes = []
         self.sweep_nodes = []
         self.z_group_recorder = []
         # quad elements flag
         self.quad_ele = quad_ele
-        self.max_grid_dim = None  #
-        # rigid link parameters - to be assigned to beam elements representing longitudinal beams
+        # get custom rigid link parameters - for future feature of beam_link model
         self.rigid_dist_y = kwargs.get("rigid_dist_y")
         self.rigid_dist_z = kwargs.get("rigid_dist_z")
         self.rigid_dist_x = kwargs.get("rigid_dist_x")
+
+        # multi span feature
+        self.multi_span_dist_list = kwargs.get("multi_span_dist_list", [self.long_dim])
+        self.multi_span_num_points = kwargs.get(
+            "multi_span_num_points",
+            [self.num_trans_beam for a in self.multi_span_dist_list],
+        )
+        self.continuous = kwargs.get("continuous", True)
+        self.stitch_slab_elements = kwargs.get("stitch_slab_elements", False)
+        self.non_cont_spacing_x = kwargs.get("non_cont_spacing_x", None)
+        # init storing var for multi span
+        self.support_points_x = [
+            sum(self.multi_span_dist_list[:n])
+            for n in range(len(self.multi_span_dist_list) + 1)
+        ]
+        self.multi_span_control_point_list = []
+        self.non_cont_support_points_x = [self.support_points_x[0]]
+        # check inputs for multi span feature
+        if len(self.multi_span_dist_list) == 1 and not self.continuous:
+            raise Exception(
+                "Combination of multi_span_dist_list and non continuous option not valid:"
+                "Hint - use only either (1) Continuous for multi span or (2) have more than"
+                "one multi_span_dist in list and non continuous option"
+            )
+
+        # if non continuous is specified, populate new variable non_cont_suport_points_x for later usage
+        if not self.continuous and self.non_cont_spacing_x:
+            for i, x_point in enumerate(self.support_points_x):
+                if (
+                    i > 0 and i != len(self.support_points_x) - 1
+                ):  # if not the first and last element, split and store into
+                    # non_cont_support_points_x
+                    left = x_point - self.non_cont_spacing_x
+                    right = x_point + self.non_cont_spacing_x
+                    self.non_cont_support_points_x.append(left)
[…]
