### 2022-04-19 87ff86f
docs: first commit of rework docs and examples * add jupyter notebooks of examples. * rework Examples.rst as landing page for ipynb examples. * add example for multi-span feature * modify conf.py for landing page feature.
