+++ docs/source/conf.py
@@ -41,12 +41,20 @@ extensions = [
     "sphinx.ext.autosummary",
     "sphinx.ext.coverage",
     "sphinx.ext.githubpages",
-    "recommonmark",
+    #"recommonmark",
+    "nbsphinx",
 ]
 
 autodoc_member_order = "bysource"
-autoclass_content = "both"
+autoclass_content = "both" # Add __init__ doc (ie. params) to class summaries
 autosummary_generate = True
+html_show_sourcelink = (
+    False  # Remove 'view source code' from top of page (for html, not python)
+)
+autodoc_inherit_docstrings = True  # If no docstring, inherit from base class
+set_type_checking_flag = True  # Enable 'expensive' imports for sphinx_autodoc_typehints
+nbsphinx_allow_errors = True  # Continue through Jupyter errors
+add_module_names = False  # Remove namespaces from class/method signatures
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
