+++ docs/source/rst/Examples.rst
@@ -1,646 +1,13 @@
 ========================
 Examples
 ========================
-Here are some examples of what you can do with *ospgrillage*.
 
-.. note::
 
-    All code herein are based on latest version 0.1.1
+.. toctree::
+   :maxdepth: 1
 
-Super T bridge model
-------------------------------------------------------------
-This example reproduces the grillage model of a super-T bridge from Caprani et al. (2017). In that study,
-the commercial software LUSAS is used to create the grillage model. Figure 1 shows the Super-T deck cross-section.
+   ../notebooks/Super_T_bridge_model
+   ../notebooks/Super_T_bridge_model_shell_beam
+   ../notebooks/mesh_feature
+   ../../../examples/workshop/super-t_example_final
 
-..  figure:: ../../_images/example_cross_section.PNG
-    :align: center
-    :scale: 25 %
-
-    Figure 1: Super-T deck cross section (after Caprani et al., 2017).
-
-
-
-Creating the grillage
-^^^^^^^^^^^^^^^^^^^^^^^^
-
-For this example, the five super-T beams and two edge beams (parapets) of Figure 1 are being modelled similar to Caprani et al. (2017)
-- with the number of longitudinal and transverse grid lines being 7 and 11 respectively.
-
-.. code-block:: python
-
-    import numpy as np
-    import ospgrillage as og
-
-    # Adopted units: N and m
-    kilo = 1e3
-    milli = 1e-3
-    N = 1
-    m = 1
-    mm = milli * m
-    m2 = m ** 2
-    m3 = m ** 3
-    m4 = m ** 4
-    kN = kilo * N
-    MPa = N / ((mm) ** 2)
-    GPa = kilo * MPa
-
-    # define material
-    concrete = og.create_material(material="concrete", code="AS5100-2017", grade="65MPa")
-
-    # define sections (parameters from LUSAS model)
-    edge_longitudinal_section = og.create_section(
-        A=0.934 * m2,
-        J=0.1857 * m3,
-        Iz=0.3478 * m4,
-        Iy=0.213602 * m4,
-        Az=0.444795 * m2,
-        Ay=0.258704 * m2,
-    )
-
-    longitudinal_section = og.create_section(
-        A=1.025 * m2,
-        J=0.1878 * m3,
-        Iz=0.3694 * m4,
-        Iy=0.113887e-3 * m4,
-        Az=0.0371929 * m2,
-        Ay=0.0371902 * m2,
-    )
-
-    transverse_section = og.create_section(
-        A=0.504 * m2,
-        J=5.22303e-3 * m3,
-        Iy=0.32928 * m4,
-        Iz=1.3608e-3 * m4,
-        Ay=0.42 * m2,
-        Az=0.42 * m2,
-    )
-
-    end_transverse_section = og.create_section(
-        A=0.504 / 2 * m2,
-        J=2.5012e-3 * m3,
-        Iy=0.04116 * m4,
-        Iz=0.6804e-3 * m4,
-        Ay=0.21 * m2,
-        Az=0.21 * m2,
-    )
-
-    # define grillage members
-    longitudinal_beam = og.create_member(section=longitudinal_section, material=concrete)
-    edge_longitudinal_beam = og.create_member(
-        section=edge_longitudinal_section, material=concrete
-    )
-    transverse_slab = og.create_member(section=transverse_section, material=concrete)
-    end_transverse_slab = og.create_member(
-        section=end_transverse_section, material=concrete
-    )
-
-    # parameters of bridge grillage
-    L = 33.5 * m  # span
-    w = 11.565 * m  # width
-    n_l = 7  # number of longitudinal members
-    n_t = 11  # number of transverse members
[…]
