+++ tests/test_osp_grillage.py
@@ -58,7 +58,7 @@ def bridge_model_42_negative(ref_bridge_properties):
         skew=-42,
         num_long_grid=7,
         num_trans_grid=5,
-        edge_beam_dist=1.05,
+        edge_beam_dist=0.5,
         mesh_type="Ortho",
     )
 
@@ -145,11 +145,11 @@ def shell_link_bridge(ref_bridge_properties):
         bridge_name="shelllink_10m",
         long_dim=33.5,
         width=11.565,
-        skew=0,
+        skew=20,
         num_long_grid=7,
         num_trans_grid=11,
         edge_beam_dist=1,
-        mesh_type="Orth",
+        mesh_type="Oblique",
         model_type="shell_beam",
         max_mesh_size_z=1,
         max_mesh_size_x=1,
@@ -173,7 +173,7 @@ def bridge_model_42_negative_custom_spacing(ref_bridge_properties):
     # define material
     I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
 
-    # construct grillage model
+    # construct grillage model - here without specifying edge distance
     example_bridge = og.create_grillage(
         bridge_name="SuperT_10m",
         long_dim=10,
@@ -181,9 +181,7 @@ def bridge_model_42_negative_custom_spacing(ref_bridge_properties):
         skew=-42,
         num_long_grid=7,
         num_trans_grid=5,
-        edge_beam_dist=1.05,
         mesh_type="Ortho",
-        ext_to_int_dist=1,
     )
 
     # set grillage member to element groups of grillage model
@@ -227,9 +225,83 @@ def test_create_shell_link_model(shell_link_bridge):
     assert og.ops.getNodeTags()
 
 
-# test creating default beam model with customize ext to int beam spacings
+# test creating default beam model without specifying edge beam distance
 def test_ext_to_int_spacing(bridge_model_42_negative_custom_spacing):
     example_bridge = bridge_model_42_negative_custom_spacing
     og.opsv.plot_model(az_el=(-90, 0), element_labels=0)
     og.plt.show()
-    assert all(example_bridge.Mesh_obj.noz == [0, 1.05, 2.05, 3.5, 4.95, 5.95, 7])
+    assert all(example_bridge.Mesh_obj.noz == [0., 1., 2.25, 3.5, 4.75, 6., 7.])
+
+
+def test_custom_beam_spacing_points(ref_bridge_properties):
+    # this test checks if the grillage is correctly created when a customized grillage spacing is provided to
+    # create_grillage function
+
+    # create a grillage with
+    custom_spacing = [1, 2, 1, 1, 2]  # first spacing starts at z = 0 direction
+
+    # define material
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # construct grillage model
+    example_bridge = og.create_grillage(
+        bridge_name="SuperT_10m",
+        long_dim=10,
+        width=7,
+        skew=0,
+        num_trans_grid=5,
+        mesh_type="Ortho",
+        beam_z_spacing=custom_spacing
+    )
+
+    # set grillage member to element groups of grillage model
+    example_bridge.set_member(I_beam, member="interior_main_beam")
+    example_bridge.set_member(exterior_I_beam, member="exterior_main_beam_1")
+    example_bridge.set_member(exterior_I_beam, member="exterior_main_beam_2")
+    example_bridge.set_member(exterior_I_beam, member="edge_beam")
+    example_bridge.set_member(slab, member="transverse_slab")
+    example_bridge.set_member(exterior_I_beam, member="start_edge")
[…]
