+++ ospgrillage/mesh.py
@@ -58,8 +58,6 @@ class Mesh:
         num_long_beam,
         skew_1,
         skew_2,
-        ext_to_int_a,
-        ext_to_int_b,
         pt1=Point(0, 0, 0),
         pt2=Point(0, 0, 0),
         pt3=None,
@@ -78,8 +76,6 @@ class Mesh:
         self.trans_dim = trans_dim
         self.edge_width_a = edge_dist_a
         self.edge_width_b = edge_dist_b
-        self.ext_to_int_a = ext_to_int_a
-        self.ext_to_int_b = ext_to_int_b
         self.width = width
         self.num_trans_beam = num_trans_beam
         self.num_long_beam = num_long_beam
@@ -146,23 +142,26 @@ class Mesh:
         self.rigid_dist_y = kwargs.get("rigid_dist_y")
         self.rigid_dist_z = kwargs.get("rigid_dist_z")
         self.rigid_dist_x = kwargs.get("rigid_dist_x")
-
-        # multi span feature
+        # ---------------------------------------------------------------------------------------------
+        # var pertaining multi span feature
+        # list containing length (x) for each nth span, default creates a list of single element based on long_dim
         self.multi_span_dist_list = kwargs.get("multi_span_dist_list", [self.long_dim])
+        # list of number of transverse beams in each span, default set num_trans_beam to each span if not provided
         self.multi_span_num_points = kwargs.get(
             "multi_span_num_points",
             [self.num_trans_beam for a in self.multi_span_dist_list],
         )
-        self.continuous = kwargs.get("continuous", True)
-        self.stitch_slab_elements = kwargs.get("stitch_slab_elements", False)
-        self.non_cont_spacing_x = kwargs.get("non_cont_spacing_x", None)
-        # init storing var for multi span
-        self.support_points_x = [
+
+        self.continuous = kwargs.get("continuous", True)  # checks if multi span meshes are linked or split
+        self.stitch_element_spacing_x = kwargs.get("non_cont_spacing_x", None)  # spacing between mesh of each spans
+
+        # init storing vars for multi span
+        self.mesh_edge_x_positions = [
             sum(self.multi_span_dist_list[:n])
             for n in range(len(self.multi_span_dist_list) + 1)
-        ]
+        ]  # list of support points in x dir of mesh
         self.multi_span_control_point_list = []
-        self.non_cont_support_points_x = [self.support_points_x[0]]
+        self.mesh_edge_x_positions_non_cont = [self.mesh_edge_x_positions[0]]
         # check inputs for multi span feature
         if len(self.multi_span_dist_list) == 1 and not self.continuous:
             raise Exception(
@@ -171,23 +170,29 @@ class Mesh:
                 "one multi_span_dist in list and non continuous option"
             )
 
-        # if non continuous is specified, populate new variable non_cont_suport_points_x for later usage
-        if not self.continuous and self.non_cont_spacing_x:
-            for i, x_point in enumerate(self.support_points_x):
+        # preprocess incorporate stitch element spacings to multi span edge points x
+        if not self.continuous and self.stitch_element_spacing_x:
+            for i, x_point in enumerate(self.mesh_edge_x_positions):
+                # search intermediate support points
                 if (
-                    i > 0 and i != len(self.support_points_x) - 1
-                ):  # if not the first and last element, split and store into
-                    # non_cont_support_points_x
-                    left = x_point - self.non_cont_spacing_x
-                    right = x_point + self.non_cont_spacing_x
-                    self.non_cont_support_points_x.append(left)
-                    self.non_cont_support_points_x.append(right)
-            self.non_cont_support_points_x.append(self.support_points_x[-1])
+                    i > 0 and i != len(self.mesh_edge_x_positions) - 1
[…]
