### 2022-05-10 f7e1435
docs: squash commit and resolve conflict * rebuild sphinx local doc files * restructure docs directory for nbsphinx
