+++ docs/build/html/_sources/rst/Running_analysis.rst.txt
@@ -1,6 +1,6 @@
-========================
+===================
 Performing analysis
-========================
+===================
 
 *ospgrillage* contains a load module which wraps `OpenSeesPy` commands to perform load analysis.
 
@@ -12,7 +12,7 @@ For all example code in this page, *ospgrillage* is imported as ``og``
 
 
 Load analysis workflow
-------------------------
+----------------------
 
 Figure 1 shows the flowchart for the load module of *ospgrillage*.
 
@@ -24,13 +24,15 @@ Figure 1 shows the flowchart for the load module of *ospgrillage*.
 
 
 Defining loads
-------------------------
+--------------
 
-Loads are created with the interface function :func:`~ospgrillage.load.create_load`. Users pass argument for `type=` to specify the load type.
+Loads are created with the interface function :func:`~ospgrillage.load.create_load`.
+Users pass argument for `type=` to specify the load type.
 Available loads types include `Point`_, `Line`_, and `Patch`_ loads.
 
-Each load type requires user to specify its load point(s). This is achieved by :func:`~ospgrillage.load.create_load_vertex` function. This function creates
-a `LoadPoint(x,y,z,p)` where `x`,`y`,`z` are the coordinates of the load point and `p` is the magnitude of the vertical loading.
+Each load type requires user to specify its load point(s).
+This is achieved by :func:`~ospgrillage.load.create_load_vertex` function.
+This function creates a `LoadPoint(x,y,z,p)` where `x`,`y`,`z` are the coordinates of the load point and `p` is the magnitude of the vertical loading.
 Note, `p` is a unit magnitude which is interpreted differently based on the load type - this will be later explained.
 By default, `y` is `0` i.e. the grillage model plane.
 
@@ -39,20 +41,24 @@ By default, `y` is `0` i.e. the grillage model plane.
     point_load_location = og.create_load_vertices(x=5, z=2, p=20)  # create load point
 
 
-Depending on the load type, a minimum number of LoadPoint namedTuple are required.
-These are set to each load type's `point#=` variable for the load type's coordinate system,
-where # is a digit from 1 to 9.
+Depending on the load type, a minimum number of ``LoadPoint`` `namedTuple` are required.
+These are set to each load type's `point#=` variable for the load type's coordinate system, where # is a digit from 1 to 9.
+Following sections will further elaborate the load vertices for each load type.
 
 Loads are generally defined in the global coordinate system with respect to the created grillage model.
 However, a user-defined local coordinate system is required when defining `Compound load`_ later on.
 
+
+
 Nodal loads
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^
 
-Nodal loads are defined using :func:`~ospgrillage.load.create_load`, specifying ``type= "nodal"``. There are six degrees-of-freedom (DOFs) for
-acting loads in each node. Nodal loads do not require a load vertex, instead it requires a `NodalForce(Fx,Fy,Fz,Mx,My,Mz)` namedtuple.
+Nodal loads are load applied directly onto nodes of grillage model.
+Nodal loads are defined using :func:`~ospgrillage.load.create_load`, specifying ``type= "nodal"``.
+There are six degrees-of-freedom (DOFs) for acting loads in each node of the grillage model.
 
-The following example creates a `NodalFroce` namedtuple and a nodal load on Node 13 of a model, with 10 unit force in both transverse X and Y directions.
+Nodal loads do not require a load vertex, instead it requires a `NodalForce(Fx,Fy,Fz,Mx,My,Mz)` `namedtuple`.
+The following example creates a `NodalForce` `namedtuple` and a nodal load on Node 13 of a model, with 10 unit force in both transverse X and Y directions.
 
 .. code-block:: python
 
[…]
