+++ .github/workflows/lint.yml
@@ -19,7 +19,7 @@ jobs:
           
       - name: Create Pull Request
         if: steps.action_black.outputs.is_formatted == 'true'
-        uses: peter-evans/create-pull-request@v4.0.1
+        uses: peter-evans/create-pull-request@v4.0.3
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           title: "Format Python code with psf/black push"
