+++ docs/requirements.txt
@@ -1,22 +1,22 @@
-numpy~=1.19.2
-openseespy~=3.3.0.0
-matplotlib~=3.3.2
-pytest~=6.1.1
-pandas~=1.3.2
-setuptools~=49.2.1
-scipy~=1.5.2
-xarray~=0.18.2
-opsvis~=0.95.5
-openseespyvis~=0.0.6
-pydata-sphinx-theme==0.7.1
-recommonmark==0.7.1
-Sphinx==4.0.2
+numpy
+openseespy
+matplotlib
+pytest>=6.1.1
+pandas
+setuptools
+scipy
+xarray
+opsvis
+openseespyvis
+pydata-sphinx-theme
+recommonmark
+Sphinx
 nbsphinx
-sphinx-rtd-theme==0.5.1
-sphinxcontrib-applehelp==1.0.2
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==1.0.3
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.4
+sphinx-rtd-theme
+sphinxcontrib-applehelp
+sphinxcontrib-devhelp
+sphinxcontrib-htmlhelp
+sphinxcontrib-jsmath
+sphinxcontrib-qthelp
+sphinxcontrib-serializinghtml
 pandoc
