### 2022-05-18 #4177 @dbrizard
Hi @prashjha, yes it is in my area of interest and I would be interested in reviewing this software...
... but I have been struggling with the installation of the module for more than one hour without success because I am running Python 3.6.8 and the module requires 3.8. So I need to install Python 3.8 (and get familiar with virtual environments for the occasion, which is new to me).
So I need to be sure I am able to import the module and run the code before accepting.
