### 2022-05-20 506bdfe
refactor: major refactor part 1 * material module refactor * refactor shell model's set_member() function
