+++ ospgrillage/material.py
@@ -39,20 +39,19 @@ def create_material(**kwargs):
 class Material:
     """
     This class stores information and provides methods to parse input material properties into ``OpenSeesPy`` Material
-    commands. ``OpenSeesPy`` has two types of material objects, namely UniaxialMaterial and NDMaterial.
+    objects.
 
     `Here <https://openseespydoc.readthedocs.io/en/latest/src/uniaxialMaterial.html>`_ are the information about OpenSees
     Material objects.
 
     As *ospgrillage* is mainly intended for bridge decks , concrete and steel makes up the primary
-    materials. In turn, UniAxialMaterial object of ``OpenSeesPy`` is wrapped and used by Material class since it contains
+    materials. In turn, *ospgrillage* wraps the UniAxialMaterial object of ``OpenSeesPy`` since it contains the primary
     options for Concrete and Steel.
 
-    The Material class also allow users to create codified material properties (e.g. AS5100).
-    These material properties are stored in a material library file (mat_lib.json) for which is to the users to pass in
-    the appropriate keyword arguments to create the code defined materials.
-
-    Additionally, Material class provides method to create the mat_lib.json file, if file is not found in directory.
+    The Material class also allow users to utilize codified material properties (e.g. AS5100, ASHTOO).
+    These material properties are stored in a material library file (mat_lib.json). Users are required to pass in
+    the appropriate keyword arguments (dict keys) to select the desirable codified materials. Users may access
+    mat_lib.json. Additionally, this class is able create the mat_lib.json file and pass it through
 
     .. note::
 
@@ -61,14 +60,13 @@ class Material:
 
     def __init__(self, **kwargs):
         """
-        The constructor of Material takes in three types of keyword arguments:
+        The constructor of Material takes in three types of inputs:
 
-        #. Keyword for looking up the *ospgrillage* material library i.e. mat_lib.json.
+        #. Keywords for looking up the *ospgrillage* material library i.e. mat_lib.json.
         #. General material properties - such as E, and G
-        #. Material arguments of ```OpenSeesPy```. E.g. Opensees's Steel01 material takes isotropic hardening parameters a1
-           to a4.
+        #. Specific material arguments of ``OpenSeesPy``.
 
-        The following keywords are for item (1):
+        For (1), the following keywords are **required**:
 
         :keyword:
 
@@ -76,25 +74,23 @@ class Material:
         * type (`str`): Either "concrete" or "steel"
         * grade(`str`): Grade of material according to code
 
-        The following keywords are examples of general material properties:
+        For (2), the minimum required material properties are:
 
         :keyword:
 
         * E (`float`): Elastic modulus
         * G (`float`): Shear modulus
+        * v (`float`): Poisson's ratio
+        * rho (`float`): Density
 
+        For (3), refer to `OpenSeesPy <https://openseespydoc.readthedocs.io/en/latest/src/uniaxialMaterial.html>`_.
 
-        For developers wishing to add more material properties:
-
-        #. if the material is a codified material, modify mat_lib.json file by adding the material following its json format.
-        #. if the material is a ``OpenSees`` Material that wasn't added previously, first add its properties under :func:`get_mat_args`
-            function. Then check the commands in :func:`_write_material` of :class:`OspGrillage` class.
 
         """
         # Instantiate variables
-        self.mat_type = None  # this is the os convention for materials e.g. Concrete01
+        self.ops_mat_type = kwargs.get("ops_mat_type", None)  # this is the os convention for materials e.g. Concrete01
         self.op_mat_arg = None  # arguments according to OpenSeesPy
-        self.units = None  # default SI units
[…]
