+++ tests/test_osp_grillage.py
@@ -5,7 +5,6 @@ import sys, os
 
 sys.path.insert(0, os.path.abspath("../"))
 
-
 # Fixtures
 @pytest.fixture
 def ref_bridge_properties():
@@ -161,6 +160,7 @@ def shell_link_bridge(ref_bridge_properties):
     example_bridge.set_member(I_beam, member="interior_main_beam")
     example_bridge.set_member(I_beam, member="exterior_main_beam_1")
     example_bridge.set_member(I_beam, member="exterior_main_beam_2")
+
     # set shell
     example_bridge.set_shell_members(slab_shell)
 
@@ -345,3 +345,153 @@ def test_inputs_custom_spacings_on_ortho_mesh(ref_bridge_properties):
         example_bridge.create_osp_model(pyfile=False)
         og.opsv.plot_model(az_el=(-90, 0), element_labels=0)
         og.plt.show()
+
+
+def test_multispan_feature(ref_bridge_properties):
+    # test multispan feature
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # Adopted units: N and m
+    kilo = 1e3
+    milli = 1e-3
+    N = 1
+    m = 1
+    mm = milli * m
+    m2 = m ** 2
+    m3 = m ** 3
+    m4 = m ** 4
+    kN = kilo * N
+    MPa = N / ((mm) ** 2)
+    GPa = kilo * MPa
+
+    # parameters of bridge grillage
+    L = 33.5 * m  # span
+    w = 11.565 * m  # width
+    n_l = 7  # number of longitudinal members
+    n_t = 11  # number of transverse members
+    edge_dist = 1.05 * m  # distance between edge beam and first exterior beam
+    bridge_name = "multi span showcase"
+    angle = 20  # degree
+    mesh_type = "Oblique"
+
+    # multispan specific vars
+    spans = [9 * m, 12 * m, 9 * m]
+    nl_multi = [20, 10, 20]
+    stich_slab_x_spacing = 0.5 * m
+
+    variant_one_model = og.create_grillage(
+        bridge_name=bridge_name,
+        long_dim=L,
+        width=w,
+        skew=angle,
+        num_long_grid=n_l,
+        num_trans_grid=n_t,
+        edge_beam_dist=edge_dist,
+        mesh_type=mesh_type,
+        multi_span_dist_list=spans,
+        multi_span_num_points=nl_multi,
+        continuous=False,
+        non_cont_spacing_x=stich_slab_x_spacing,
+    )
+
+    # assign grillage member to element groups of grillage model
+    variant_one_model.set_member(I_beam, member="interior_main_beam")
+    variant_one_model.set_member(I_beam, member="exterior_main_beam_1")
+    variant_one_model.set_member(I_beam, member="exterior_main_beam_2")
+    variant_one_model.set_member(exterior_I_beam, member="edge_beam")
+    variant_one_model.set_member(slab, member="transverse_slab")
+    variant_one_model.set_member(slab, member="start_edge")
+    variant_one_model.set_member(slab, member="end_edge")
+    # variant_one_model.set_member(stich_slab, member="stitch_elements")
+
+    variant_one_model.create_osp_model(pyfile=False)
+    og.opsv.plot_model(element_labels=0, az_el=(-90, 0))  # plotting using ops_vis
+    og.plt.show()
+
+
+def test_multispan_feat_shell(ref_bridge_properties):
+    # test multispan feature compatibility with shell model
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # Adopted units: N and m
+    kilo = 1e3
+    milli = 1e-3
+    N = 1
+    m = 1
+    mm = milli * m
[…]
