### 2022-06-07 5e16a84
refactor: major refactor part 2 * first version of curve mesh feature * added new tests for every module * improve docstrings * refactor vars
