+++ tests/test_postprocessing.py
@@ -74,8 +74,51 @@ def bridge_model_42_negative(ref_bridge_properties):
     example_bridge.create_osp_model(pyfile=pyfile)
     return example_bridge
 
+@pytest.fixture
+def shell_link_bridge(ref_bridge_properties):
+    # reference bridge 10m long, 7m wide with common skew angle at both ends
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # create material of slab shell
+    slab_shell_mat = og.create_material(
+        material="concrete", code="AS5100-2017", grade="50MPa", rho=2400
+    )
+
+    # create section of slab shell
+    slab_shell_section = og.create_section(h=0.2)
+    # shell elements for slab
+    slab_shell = og.create_member(section=slab_shell_section, material=slab_shell_mat)
+
+    # construct grillage model
+    example_bridge = og.create_grillage(
+        bridge_name="shelllink_10m",
+        long_dim=10,
+        width=7,
+        skew=0,
+        num_long_grid=6,
+        num_trans_grid=11,
+        edge_beam_dist=1,
+        mesh_type="Orth",
+        model_type="shell_beam",
+        max_mesh_size_z=1,
+        max_mesh_size_x=1,
+        offset_beam_y_dist=0.499,
+        beam_width=0.89,
+    )
+
+    # set beams
+    example_bridge.set_member(I_beam, member="interior_main_beam")
+    example_bridge.set_member(I_beam, member="exterior_main_beam_1")
+    example_bridge.set_member(I_beam, member="exterior_main_beam_2")
+    # set shell
+    example_bridge.set_shell_members(slab_shell)
+
+    example_bridge.create_osp_model(pyfile=False)
+    return example_bridge
+
 
 def test_envelope(bridge_model_42_negative):
+    # test functionality of Envelope class and output
     og.ops.wipeAnalysis()
     example_bridge = bridge_model_42_negative
 
@@ -166,3 +209,114 @@ def test_envelope(bridge_model_42_negative):
     )
     # print(comb_results)
     print(og.ops.nodeDisp(25)[1])
+
+
+def test_plot_force(bridge_model_42_negative):
+    # test functionality of plot_force and its output
+    og.ops.wipeAnalysis()
+    example_bridge = bridge_model_42_negative
+
+    # create reference line load
+    p = 10000
+    p2 = 20000
+    p3 = 30000  # duplicate of 2nd but with different mag
+    barrierpoint_1 = og.create_load_vertex(x=5, z=1, p=p)
+    barrierpoint_2 = og.create_load_vertex(x=10, z=7, p=p)
+    barrierpoint_3 = og.create_load_vertex(x=10, z=2, p=p2)
+    barrierpoint_4 = og.create_load_vertex(x=5, z=5, p=p2)
+    barrierpoint_5 = og.create_load_vertex(x=10, z=2, p=p3)
+    barrierpoint_6 = og.create_load_vertex(x=5, z=5, p=p3)
+    # add moving load case
+    front_wheel = og.PointLoad(
+        name="front wheel", point1=og.LoadPoint(2, 0, 2, 50)
+    )  # Single point load 50 N
+    Barrier = og.create_load(
+        loadtype="line",
+        name="Barrier curb load",
+        point1=barrierpoint_1,
+        point2=barrierpoint_2,
+    )
+    Barrier2 = og.create_load(
+        loadtype="line", name="Barrieload", point1=barrierpoint_3, point2=barrierpoint_4
+    )
+    Barrier3 = og.create_load(
+        loadtype="line", name="Barrieload", point1=barrierpoint_5, point2=barrierpoint_6
+    )
+    Patch1 = og.create_load(
+        loadtype="patch",
+        point1=barrierpoint_1,
[…]
