+++ ospgrillage/mesh.py
@@ -15,7 +15,9 @@ from collections import namedtuple
 
 def create_point(**kwargs):
     """
-    User interface function to create a point named tuple
+    User interface function to create a point named tuple - this is used in defining positions such as loads for
+    example.
+
     :keyword:
 
     * x (`float` or `int`): x coordinate
@@ -37,7 +39,7 @@ Point = namedtuple("Point", ["x", "y", "z"])
 class Mesh:
     """
     Base class for mesh class. The class holds information pertaining the mesh group such as element connectivity and nodes
-    of the mesh object. Positional arguments are handled by OspGrillage class.
+    of the mesh object. Positional arguments are handled by :class:`~ospgrillage.osp_grillage.OspGrillage` class.
 
     .. note::
 
@@ -121,29 +123,28 @@ class Mesh:
         self.curve_center = []
         self.curve_radius = []
         # init line / circle equation variables
-        self.m = 0
-        self.c = 0
-        self.r = 0
-        self.R = 0
+        self.m = 0  # gradient of linear line eq for sweep path
+        self.c = 0  # local y (global x) interp of line for sweep path
+        self.r = 0  # radius of arch for swee line
+        self.R = 0  # centre of circle
         self.d = None  # list of circle centre and circle radius [ c, r]
-        # meshing properties
+        # variables for meshing
         if mesh_origin is None:
             self.mesh_origin = [0, 0, 0]
         else:
             self.mesh_origin = mesh_origin  # default origin
-        # init meshing variables
         self.first_connecting_region_nodes = []
         self.end_connecting_region_nodes = []
         self.sweep_nodes = []
         self.z_group_recorder = []
         # quad elements flag
-        self.quad_ele = quad_ele
+        self.quad_ele = quad_ele  # bool
         # get custom rigid link parameters - for future feature of beam_link model
         self.rigid_dist_y = kwargs.get("rigid_dist_y")
         self.rigid_dist_z = kwargs.get("rigid_dist_z")
         self.rigid_dist_x = kwargs.get("rigid_dist_x")
         # ---------------------------------------------------------------------------------------------
-        # var pertaining multi span feature
+        # vars for multi span feature
         # list containing length (x) for each nth span, default creates a list of single element based on long_dim
         self.multi_span_dist_list = kwargs.get("multi_span_dist_list", [self.long_dim])
         # list of number of transverse beams in each span, default set num_trans_beam to each span if not provided
@@ -159,7 +160,7 @@ class Mesh:
             "non_cont_spacing_x", None
         )  # spacing between mesh of each spans
 
-        # init storing vars for multi span
+        # to store positions and points of spans
         self.mesh_edge_x_positions = [
             sum(self.multi_span_dist_list[:n])
             for n in range(len(self.multi_span_dist_list) + 1)
@@ -174,7 +175,9 @@ class Mesh:
                 "one multi_span_dist in list and non continuous option"
             )
 
-        # preprocess incorporate stitch element spacings to multi span edge points x
+        # preprocess to incorporate stitch element spacings to multi span edge points x
+        # node at support point is split into two (equally in both sides in x directions) nodes, which are assigned
+        # as supports later on. Stitch elements are subsequently assigned between the two new nodes.
         if not self.continuous and self.stitch_element_spacing_x:
             for i, x_point in enumerate(self.mesh_edge_x_positions):
                 # search intermediate support points
@@ -186,12 +189,12 @@ class Mesh:
                     self.mesh_edge_x_positions_non_cont.append(right)
[…]
