### 2022-06-08 c9b2259
refactor: major refactor part 3 * v2 curve mesh feature * add tests for curve mesh and static methods
