+++ ospgrillage/mesh.py
@@ -241,6 +241,7 @@ class Mesh:
             edge_angle=self.skew_1,
             num_long_beam=self.num_long_beam,
             model_plane_y=self.y_elevation,
+            sweep_path = self.sweep_path,
             **kwargs
         )
         # intermediate construction lines for
@@ -260,6 +261,7 @@ class Mesh:
             edge_angle=self.skew_2,
             num_long_beam=self.num_long_beam,
             model_plane_y=self.y_elevation,
+            sweep_path=self.sweep_path,
             **kwargs
         )
         # ------------------------------------------------------------------------------------------
@@ -304,14 +306,19 @@ class Mesh:
         self._mesh_grillage()
 
     def _create_custom_transverse_spacings(self):
+        # populate nox with custom spacing provided by users
         self.nox = [0]
         for x_dist in self.transverse_mbr_x_spacing_list:
             self.nox.append(self.nox[-1] + x_dist)
 
     def _create_transverse_spacings(self):
-        # loop through the first to nth span
+        # loop through each span group
         for i, num_x_point in enumerate(self.multi_span_num_points):
             # TODO check if straight line or equation of arc
+            if self.sweep_path.curve_path:  # if curve
+                pass
+            else:
+                pass
             # arc
 
             # straight line
@@ -365,6 +372,7 @@ class Mesh:
         return EdgeControlLine(**kwargs)
 
     # ------------------------------------------------------------------------------------------
+
     # main meshing algorithms (straight mesh)
     def _fixed_sweep_node_meshing(self):
         assigned_node_tag = []
@@ -520,6 +528,9 @@ class Mesh:
         if np.abs(self.skew_1 + self.zeta) < self.skew_threshold[0]:
             # if angle less than threshold, assign nodes of edge member as it is
             current_sweep_nodes = self.start_edge_line.node_list
+            # if curve mesh, rotate the edge sweep nodes
+            current_sweep_nodes = self._rotate_edge_sweep_nodes(current_sweep_nodes)
+
             for (z_count_int, nodes) in enumerate(current_sweep_nodes):
                 x_inc = start_point_x
                 z_inc = start_point_z
@@ -694,7 +705,12 @@ class Mesh:
         end_point_z = self.sweep_path.get_line_function(end_point_x)
         if np.abs(self.skew_2 + self.zeta) < self.skew_threshold[0]:
             # if angle less than threshold, assign nodes of edge member as it is
-            current_sweep_nodes = self.end_edge_line.node_list
+            current_sweep_nodes = self.start_edge_line.node_list
+            edge_angle = self.sweep_path.get_cartesian_angle(x=end_point_x)
+            # if curve mesh, rotate the edge sweep nodes
+            #current_sweep_nodes = self._rotate_sweep_nodes(-edge_angle)
+            current_sweep_nodes = self._rotate_edge_sweep_nodes(current_sweep_nodes,angle=-edge_angle)
+
             for (z_count_int, nodes) in enumerate(current_sweep_nodes):
                 x_inc = 0  # end_point_x
                 z_inc = 0  # end_point_z
@@ -750,6 +766,7 @@ class Mesh:
                 m_prime, phi = get_slope(
                     [ref_point_x, self.y_elevation, ref_point_z], int_point
                 )
+
                 # rotate sweep line such that parallel to m' line
                 current_sweep_nodes = self._rotate_sweep_nodes(np.pi / 2 - np.abs(phi))
                 # get z group of first node in current_sweep_nodes - for correct assignment in loop
@@ -863,17 +880,22 @@ class Mesh:
         for z_count, x in enumerate(self.uniform_region_x[1:-1]):
             # get slope, m at current point x
[…]
