+++ tests/test_static.py
@@ -34,3 +34,12 @@ def test_create_arc_equation():
 
     assert a[0] == [1, 0]
     assert a[1] == 1
+
+
+def test_rotating_points():
+    # check point rotating function
+    rotated_coord = og.rotate_point_about_point(center_x=0,center_y=-11,angle=32,point=[10,-6])
+
+    print(rotated_coord)
+    assert og.np.isclose(rotated_coord[0],5.585100198856649)
+    assert og.np.isclose(rotated_coord[1],-1.3146163850505417)
