### 2022-06-09 307de8a
refactor: major refactor part 4 * v3 curve mesh feature * fix errors from tests
