### 2022-06-13 #4404 @dbrizard
This is a detail, but the *enforcement* section of the code of conduct has not been personalised.
https://github.com/MonashSmartStructures/ospgrillage/blob/main/.github/CODE_OF_CONDUCT.md
