### 2022-06-13 #4404 @dbrizard
Two **references** should be improved, even though they are not so hard to find (because it is not possible to guess what type of reference it is from the bibliography alone):

- Almutairi et al: I suggest adding the report number and the institution;
- Yang et al: add the proceeding title and the place.
