### 2022-06-13 #4404 @dbrizard
Even if I am not in the "brigdes" field, I believe this software is a important contribution to the community.
The code is very clean, well organised, the documentation is exhaustive and with many examples which greatly facilitates the understanding and getting started phase. This software is worth being published to JOSS :+1:
