### 2022-06-28 #4404 @vsangelidakis
I am also satisfied to see a well-implemented Python package, detailed documentation and a clear manuscript. Seeing the docs hosted on github pages (rather than using a third-party domain) makes me confident about their longevity.

One comment regarding the **State of the Field** item of the reviewing TODO list above: the authors state "there is no established interface to import such models into OpenSees framework" and I completely agree that no pre-processor exists for grillage models specifically.

I would like to bring to your attention that other pre-processors do exist for OpenSees, e.g. (1) generic GUI tools such as GiD-OpenSees https://github.com/rclab-auth/gidopensees or (2) for bridges in general, such as the Bridge-Wizard (https://research-information.bris.ac.uk/en/publications/bridge-wizard-expert-system-for-finite-element-modeling-and-post-); although I could only find binaries and not the source code itself for the latter (see https://www.asextos.net/bridge-wizard.html).

I leave it to the judgement of the authors; you can either leave the manuscript as is or mention how ospgrillage is an improvement compared to existing pre-processing tools like the ones above. Either way, please let me know your intention.
