### 2022-06-28 #4404 @vsangelidakis
A small point for improvement: The documentation describes how to generate multi-span models (https://monashsmartstructures.github.io/ospgrillage/notebooks/multi_span.html), while the manuscript mentions "the roadmap allows for curved and multi-span bridge decks..". Is this feature developed/operational or ongoing? Please clarify either the manuscript or the documentation.
