### 2022-06-29 c87e239
refactor: major refactor part 5 * fist restructure of docs index.rst * refactor ipynb
