+++ paper/paper.bib
@@ -29,6 +29,8 @@
     month = {01},
     pages = {},
     title = {Seismic Assessment for Retrofitted Skewed Reinforced Concrete Bridges with Buckling Restrained Braces}
+    booktitle = {Proceedings of the 16th World Conference on Earthquake Engineering (16WCEE)}
+    address - {Santiago, Chile}
 }
 
 @article{Elhami2014,
@@ -56,6 +58,8 @@
     title={Analysis of Multi-span Bridges Using OpenSees},
     author={Abdullah Almutairi and Jinchi Lu and Ning Wang and Ahmed Elgamal},
     year={2016}
+    number={SSRP-16/01}
+    institution={Department of Structural Engineering University of California, San Diego La Jolla, California 92093-0085 }
 }
 
 @article{Benjumea,
