+++ docs/source/rst/user_guide.rst
@@ -1,0 +1,14 @@
+========================
+User Guide
+========================
+
+.. toctree::
+   :maxdepth: 1
+
+   rst/Installation
+   rst/Module description
+   rst/ModelTemplates
+   rst/mesh_feature
+   rst/Running_analysis
+   rst/PostProcessing
+   rst/Examples
