+++ docs/source/rst/mesh_feature.rst
@@ -1,0 +1,17 @@
+====================================
+Multi span feature
+====================================
+
+.. toctree::
+   :maxdepth: 1
+
+   ../notebooks/mesh_showcase
+
+====================================
+Curve mesh feature
+====================================
+
+.. toctree::
+   :maxdepth: 1
+
+   ../notebooks/multi_span
