### 2022-07-01 #4404 @dbrizard
@prashjha yes, provided the references are updated according to my suggestion
https://github.com/openjournals/joss-reviews/issues/4404#issuecomment-1153923960
otherwise everything is ok for me.
