### 2022-07-01 974183c
refactor: major refactor part 6 * local sphinx rebuild * rework doc structure * fix multi-span feature adding extra longitudinal members
