+++ tests/test_osp_grillage.py
@@ -5,6 +5,7 @@ import sys, os
 
 sys.path.insert(0, os.path.abspath("../"))
 
+
 # Fixtures
 @pytest.fixture
 def ref_bridge_properties():
@@ -371,13 +372,22 @@ def test_multispan_feature(ref_bridge_properties):
     n_t = 11  # number of transverse members
     edge_dist = 1.05 * m  # distance between edge beam and first exterior beam
     bridge_name = "multi span showcase"
-    angle = 20  # degree
+    angle = 10  # degree
     mesh_type = "Oblique"
 
     # multispan specific vars
     spans = [9 * m, 12 * m, 9 * m]
-    nl_multi = [20, 10, 20]
-    stich_slab_x_spacing = 0.5 * m
+    nl_multi = [3, 5, 10]
+    stich_slab_x_spacing = 1 * m
+    stitch_slab_section = og.create_section(
+        A=0.504 * m2,
+        J=5.22303e-3 * m3,
+        Iy=0.32928 * m4,
+        Iz=1.3608e-3 * m4,
+        Ay=0.42 * m2,
+        Az=0.42 * m2,
+    )
+    stich_slab = og.create_member(section=stitch_slab_section, material=concrete)
 
     variant_one_model = og.create_grillage(
         bridge_name=bridge_name,
@@ -390,8 +400,8 @@ def test_multispan_feature(ref_bridge_properties):
         mesh_type=mesh_type,
         multi_span_dist_list=spans,
         multi_span_num_points=nl_multi,
-        continuous=False,
-        non_cont_spacing_x=stich_slab_x_spacing,
+        continuous=True,
+        # non_cont_spacing_x=stich_slab_x_spacing,
     )
 
     # assign grillage member to element groups of grillage model
@@ -407,6 +417,8 @@ def test_multispan_feature(ref_bridge_properties):
     variant_one_model.create_osp_model(pyfile=False)
     og.opsv.plot_model(element_labels=0, az_el=(-90, 0))  # plotting using ops_vis
     og.plt.show()
+    assert all(og.np.isclose(variant_one_model.Mesh_obj.nox, [0., 4.5, 9., 12., 15., 18., 21., 22., 23., 24., 25.,
+                                                              26., 27., 28., 29., 30.]))
 
 
 def test_multispan_feat_shell(ref_bridge_properties):
@@ -470,7 +482,7 @@ def test_multispan_feat_shell(ref_bridge_properties):
     slab_shell_section = og.create_section(h=0.2)
     slab_shell = og.create_member(section=slab_shell_section, material=slab_shell_mat)
 
-    # create stitch slab conencting elements
+    # create stitch slab connecting elements
     stitch_slab_section = og.create_section(
         A=0.504 * m2,
         J=5.22303e-3 * m3,
@@ -493,8 +505,22 @@ def test_multispan_feat_shell(ref_bridge_properties):
 
     variant_one_model.create_osp_model(pyfile=False)
     og.opsplt.plot_model()
-    #og.opsv.plot_model(element_labels=0, az_el=(-90, 0))  # plotting using ops_vis
-    #og.plt.show()
+    # og.opsv.plot_model(element_labels=0, az_el=(-90, 0))  # plotting using ops_vis
+    # og.plt.show()
+    assert all(og.np.isclose(variant_one_model.Mesh_obj.nox, [0., 0.44736842, 0.89473684, 1.34210526, 1.78947368,
+                                                              2.23684211, 2.68421053, 3.13157895, 3.57894737,
+                                                              4.02631579,
+                                                              4.47368421, 4.92105263, 5.36842105, 5.81578947,
+                                                              6.26315789,
[…]
