+++ docs/source/rst/mesh_feature.rst
@@ -1,17 +1,29 @@
 ====================================
-Multi span feature
+Mesh features
 ====================================
 
+Mesh capabilities
+-----------------------------
 .. toctree::
    :maxdepth: 1
 
    ../notebooks/mesh_showcase
 
-====================================
-Curve mesh feature
-====================================
+
+Multi span feature
+-----------------------------
 
 .. toctree::
    :maxdepth: 1
 
    ../notebooks/multi_span
+
+
+
+Curve mesh feature
+-----------------------------
+
+.. toctree::
+   :maxdepth: 1
+
+   ../notebooks/curve_mesh
