+++ docs/source/rst/user_guide.rst
@@ -5,10 +5,10 @@ User Guide
 .. toctree::
    :maxdepth: 1
 
-   rst/Installation
-   rst/Module description
-   rst/ModelTemplates
-   rst/mesh_feature
-   rst/Running_analysis
-   rst/PostProcessing
-   rst/Examples
+   Installation
+   Module description
+   ModelTemplates
+   mesh_feature
+   Running_analysis
+   PostProcessing
+   Examples
