### 2022-07-07 3ca87ff
Fix: pygments for docs build
