+++ docs/requirements.txt
@@ -21,3 +21,4 @@ sphinxcontrib-qthelp
 sphinxcontrib-serializinghtml
 pandoc
 markupsafe==2.0.1
+pygments >= 2.7
