+++ _sources/notebooks/curve_mesh.ipynb.txt
@@ -1,0 +1,300 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "ff62ae04-bbda-4e1c-bbaa-84b11a07b120",
+   "metadata": {},
+   "source": [
+    "# Curve Mesh\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "fb986f85-064b-476e-818d-c88071a6033c",
+   "metadata": {
+    "tags": []
+   },
+   "source": [
+    "`ospgrillage` allows curve mesh to be implemented. In this example we will explore the steps to define a curve mesh. In short, curve mesh are defined by setting defining a radius, `R` parameter. `ospgrillage` uses the radius along with the fundamental inputs (span, `L` and width ,`w` etc) to determine the curvature of the mesh (Figure 1)."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 17,
+   "id": "2eda936e-21fa-4ca1-a69a-dd56950cb671",
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
[…]
