### 2022-07-07 #4404 @ccaprani
Hi all,
Thank you so much for your valuable time and for the useful and constructive reviews. We've now released v0.2 with completion of the curved and multi-span meshes and a major update of the documentation.
@vsangelidakis thank you - we've updated the documentation now with more extensive description of these features. Those other pre-processors mentioned all address different domains. Further, they are either aimed at commercial software, or are not open-source.
@dbrizard thanks - these are updated.
@prashjha I think that is everything but please do let me know if there is anything else to be addressed.
