### 2022-08-01 18b0ea2
Fix paper bibtex formatting
