+++ paper/paper.bib
@@ -28,9 +28,9 @@
     year = {2017},
     month = {01},
     pages = {},
-    title = {Seismic Assessment for Retrofitted Skewed Reinforced Concrete Bridges with Buckling Restrained Braces}
-    booktitle = {Proceedings of the 16th World Conference on Earthquake Engineering (16WCEE)}
-    address - {Santiago, Chile}
+    title = {Seismic Assessment for Retrofitted Skewed Reinforced Concrete Bridges with Buckling Restrained Braces},
+    booktitle = {Proceedings of the 16th World Conference on Earthquake Engineering (16WCEE)},
+    address = {Santiago, Chile}
 }
 
 @article{Elhami2014,
