### 2022-08-01 4cb3a96
FIx paper bitex #2
