+++ paper/paper.bib
@@ -57,8 +57,8 @@
 @inproceedings{Almutairi2016AnalysisOM,
     title={Analysis of Multi-span Bridges Using OpenSees},
     author={Abdullah Almutairi and Jinchi Lu and Ning Wang and Ahmed Elgamal},
-    year={2016}
-    number={SSRP-16/01}
+    year={2016},
+    number={SSRP-16/01},
     institution={Department of Structural Engineering University of California, San Diego La Jolla, California 92093-0085 }
 }
 
@@ -84,6 +84,5 @@
     doi = {10.1533/9780857099754.331},
     url = {https://www.sciencedirect.com/science/article/pii/B9781898563518500172},
     author = {Carl T.F. Ross},
-    abstract = {Publisher Summary
-    This chapter is concerned with the static and vibration analysis of orthogonal and skew grillages, and a number of such structures are analyzed by computer in the chapter. Grillages or grids are often used to cover wide areas, where the supports can, in general, only be placed around the periphery of the structure and at a few other discrete points. The stiffening effect of cross-members is used to strengthen the structure so that the lateral loads can be resisted in bending and torsion by the members of the grid in both directions. Gillage structures are of much interest in many branches of engineering, including the aircraft, shipbuilding, and the mechanical and civil engineering industries. These structures take many various forms, including hatch covers, decks, platforms, floors, etc. Much of the work that has been done on grillage structures has been based on analytical techniques and these theories have usually considered the static analysis of orthogonally stiffened rectangular grillages under uniform pressure. Whereas these theories are very useful for a large number of problems, they break down for gillages of complex shape with complex boundary conditions and load, and where the members of the grillage are made from different sections.}
+    abstract = {This chapter is concerned with the static and vibration analysis of orthogonal and skew grillages, and a number of such structures are analyzed by computer in the chapter. Grillages or grids are often used to cover wide areas, where the supports can, in general, only be placed around the periphery of the structure and at a few other discrete points. The stiffening effect of cross-members is used to strengthen the structure so that the lateral loads can be resisted in bending and torsion by the members of the grid in both directions. Gillage structures are of much interest in many branches of engineering, including the aircraft, shipbuilding, and the mechanical and civil engineering industries. These structures take many various forms, including hatch covers, decks, platforms, floors, etc. Much of the work that has been done on grillage structures has been based on analytical techniques and these theories have usually considered the static analysis of orthogonally stiffened rectangular grillages under uniform pressure. Whereas these theories are very useful for a large number of problems, they break down for gillages of complex shape with complex boundary conditions and load, and where the members of the grillage are made from different sections.}
 }
