### 2022-08-24 #4404 @prashjha
Hi, @ccaprani; the paper is well written and informative. My only suggestion is to replace `Dr` with `Dr.` in line 102 of the draft.

After you have updated the draft, can you also do (if not done already) a 'tagged' release of your code and archive the release using zenodo or other methods? Make sure that the zenodo archive's title matches this JOSS submission's title.

Once you are done, I will hand your paper to EiC for the final decision.
