### 2022-08-30 #4404 @ccaprani
Thank you @prashjha ; those final tweaks are done; v0.2.1 released, and archived at [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7034267.svg)](https://doi.org/10.5281/zenodo.7034267).

I appreciate your time on this, and hope for a decision soon. Thanks.
