### 2022-08-30 #4404 @prashjha
Hi @ccaprani, thanks. One last thing, could you change the title of [your zenodo archive](https://doi.org/10.5281/zenodo.7034267) to match the title of your JOSS paper?
