### 2022-09-01 cb06349
feat: spring support * function to define rotational spring at supports * modify material and member module to accomodate for spring (OpenSees'zeroLength) elements. * add test for analysis entailing new spring support model.
