+++ tests/test_load.py
@@ -1241,3 +1241,93 @@ def test_load_analysis_shell_multi_span(ref_bridge_properties):
     # extract results
     result = shell_bridge.get_results()
     print(result)
+
+
+def test_load_analysis_on_spring_support_single_span(ref_bridge_properties):
+    # test multispan feature
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # Adopted units: N and m
+    kilo = 1e3
+    milli = 1e-3
+    N = 1
+    m = 1
+    mm = milli * m
+    m2 = m ** 2
+    m3 = m ** 3
+    m4 = m ** 4
+    kN = kilo * N
+    MPa = N / ((mm) ** 2)
+    GPa = kilo * MPa
+
+    # parameters of bridge grillage
+    L = 33.5 * m  # span
+    w = 11.565 * m  # width
+    n_l = 7  # number of longitudinal members
+    n_t = 11  # number of transverse members
+    edge_dist = 1.05 * m  # distance between edge beam and first exterior beam
+    bridge_name = "multi span showcase"
+    angle = 10  # degree
+    mesh_type = "Oblique"
+
+    # multispan specific vars
+    spans = [9 * m, 12 * m, 9 * m]
+    nl_multi = [3, 5, 10]
+    stich_slab_x_spacing = 1 * m
+    stitch_slab_section = og.create_section(
+        A=0.504 * m2,
+        J=5.22303e-3 * m3,
+        Iy=0.32928 * m4,
+        Iz=1.3608e-3 * m4,
+        Ay=0.42 * m2,
+        Az=0.42 * m2,
+    )
+    stich_slab = og.create_member(section=stitch_slab_section, material=concrete)
+
+    variant_one_model = og.create_grillage(
+        bridge_name=bridge_name,
+        long_dim=L,
+        width=w,
+        skew=angle,
+        num_long_grid=n_l,
+        num_trans_grid=n_t,
+        edge_beam_dist=edge_dist,
+        mesh_type=mesh_type,
+        multi_span_dist_list=spans,
+        multi_span_num_points=nl_multi,
+        continuous=True,
+        #non_cont_spacing_x=stich_slab_x_spacing,
+    )
+
+    # assign grillage member to element groups of grillage model
+    variant_one_model.set_member(I_beam, member="interior_main_beam")
+    variant_one_model.set_member(I_beam, member="exterior_main_beam_1")
+    variant_one_model.set_member(I_beam, member="exterior_main_beam_2")
+    variant_one_model.set_member(exterior_I_beam, member="edge_beam")
+    variant_one_model.set_member(slab, member="transverse_slab")
+    variant_one_model.set_member(slab, member="start_edge")
+    variant_one_model.set_member(slab, member="end_edge")
+    # variant_one_model.set_member(stich_slab, member="stitch_elements")
+
+    # spring support
+    e_spring = 11e13
+    variant_one_model.set_spring_support(rotational_spring_stiffness=e_spring,edge_num=1)
+
+    variant_one_model.create_osp_model(pyfile=False)
+    og.opsplt.plot_model()
+
+    # create and add load case comprise of single point load
+    P = 20e3
+    point_load_location = og.create_load_vertex(
+        x=7.5, y=0, z=3, p=P
+    )  # about midspan of span 1
+    point_load = og.create_load(
+        name="single point", point1=point_load_location
+    )
+    point_lc = og.create_load_case(name="pointload")
+    point_lc.add_load(point_load)
+    variant_one_model.add_load_case(point_lc)
+    variant_one_model.analyze()
+
+    result = variant_one_model.get_results()
+    print(result)
