+++ ospgrillage/mesh.py
@@ -50,27 +50,27 @@ class Mesh:
     """
 
     def __init__(
-        self,
-        long_dim,
-        width,
-        trans_dim,
-        edge_dist_a,
-        edge_dist_b,
-        num_trans_beam,
-        num_long_beam,
-        skew_1,
-        skew_2,
-        pt1=Point(0, 0, 0),
-        pt2=Point(0, 0, 0),
-        pt3=None,
-        element_counter=1,
-        node_counter=1,
-        transform_counter=0,
-        global_x_grid_count=0,
-        global_edge_count=0,
-        mesh_origin=None,
-        quad_ele=False,
-        **kwargs
+            self,
+            long_dim,
+            width,
+            trans_dim,
+            edge_dist_a,
+            edge_dist_b,
+            num_trans_beam,
+            num_long_beam,
+            skew_1,
+            skew_2,
+            pt1=Point(0, 0, 0),
+            pt2=Point(0, 0, 0),
+            pt3=None,
+            element_counter=1,
+            node_counter=1,
+            transform_counter=0,
+            global_x_grid_count=0,
+            global_edge_count=0,
+            mesh_origin=None,
+            quad_ele=False,
+            **kwargs
     ):
 
         # inputs from OspGrillage required to create mesh
@@ -479,8 +479,8 @@ class Mesh:
                                 # current step is in between two span groups
                                 # check to either create
                                 if (
-                                    not self.continuous
-                                    and self.stitch_element_spacing_x
+                                        not self.continuous
+                                        and self.stitch_element_spacing_x
                                 ):
                                     # create beam element between supports
                                     x_start = self.node_spec[current_node][
@@ -525,7 +525,7 @@ class Mesh:
                 # here updates the record for previous node tag step
                 previous_node_tag = assigned_node_tag
                 if (
-                    x_inc in support_points
+                        x_inc in support_points
                 ):  # if x inc is a support roll (intermediate) set all nodes as support
                     for nodes in previous_node_tag:
                         self.edge_node_recorder.setdefault(
@@ -631,7 +631,7 @@ class Mesh:
                     sweep_nodes = current_sweep_nodes[z_count:]
                     z_group_recorder = list(range(z_group, len(current_sweep_nodes)))
                 elif 90 + self.skew_1 + self.zeta < 90:
-                    sweep_nodes = current_sweep_nodes[0 : (z_count + 1)]
+                    sweep_nodes = current_sweep_nodes[0: (z_count + 1)]
                     z_group_recorder = (
                         list(range(0, z_group + 1)) if z_group != 0 else [0]
                     )
@@ -811,7 +811,7 @@ class Mesh:
                 # check
                 # condition
                 if 90 + self.skew_2 + self.zeta > 90:
-                    sweep_nodes = current_sweep_nodes[0 : (z_count + 1)]
+                    sweep_nodes = current_sweep_nodes[0: (z_count + 1)]
                     z_group_recorder = (
                         list(range(0, z_group + 1)) if z_group != 0 else [0]
                     )
@@ -1015,7 +1015,7 @@ class Mesh:
         self.element_counter += 1
 
     def _assign_edge_trans_members(
-        self, previous_node_tag, assigned_node_tag, edge_counter
+            self, previous_node_tag, assigned_node_tag, edge_counter
     ):
         tag = self._get_geo_transform_tag([previous_node_tag, assigned_node_tag])
         self.edge_span_ele.append(
@@ -1090,9 +1090,9 @@ class Mesh:
                     [
                         np.abs(a - b)
                         for (a, b) in zip(
-                            self.node_spec[item[1]]["coordinate"],
-                            self.node_spec[item[2]]["coordinate"],
-                        )
[…]
