+++ tests/test_osp_grillage.py
@@ -643,3 +643,39 @@ def test_basic_curve_mesh(ref_bridge_properties):
             ],
         )
     )
+
+
+def test_spring_support(ref_bridge_properties):
+    I_beam, slab, exterior_I_beam, concrete = ref_bridge_properties
+
+    # construct grillage model
+    example_bridge = og.create_grillage(
+        bridge_name="SuperT_10m",
+        long_dim=10,
+        width=7,
+        skew=-42,
+        num_long_grid=7,
+        num_trans_grid=5,
+        edge_beam_dist=0.5,
+        mesh_type="Ortho",
+    )
+
+    # set grillage member to element groups of grillage model
+    example_bridge.set_member(I_beam, member="interior_main_beam")
+    example_bridge.set_member(exterior_I_beam, member="exterior_main_beam_1")
+    example_bridge.set_member(exterior_I_beam, member="exterior_main_beam_2")
+    example_bridge.set_member(exterior_I_beam, member="edge_beam")
+    example_bridge.set_member(slab, member="transverse_slab")
+    example_bridge.set_member(exterior_I_beam, member="start_edge")
+    example_bridge.set_member(exterior_I_beam, member="end_edge")
+
+    # spring support
+    e_spring = 1e9
+    #example_bridge.set_spring_support(rotational_spring_stiffness=e_spring,edge_num=0)
+    example_bridge.set_spring_support(rotational_spring_stiffness=e_spring,edge_num=1)
+
+    example_bridge.create_osp_model(pyfile=False)
+    og.opsplt.plot_model()
+
+    #print(og.ops.nodeDisp(20))
+

[…]
